I took three weather services as a basis: - Weather API - Ninjas API - Weather Stack API
In WeatherServiceImpl, I wrote the logic for working with API data, so far separately and saving the values in the test database. I also created POJO classes for data deserialization, for each API. I tested the operation of each method in the service class, everything works correctly.

+++ src/main/java/com/myproject/weatherrestapi/entity/Weather.java
@@ -19,7 +19,7 @@ public class Weather {
     private String city;
 
     @Column(name = "temperature")
-    private Integer temperature;
+    private Double temperature;
 
     @Column(name = "created_at")
     private OffsetDateTime createdAt;
@@ -56,11 +56,11 @@ public class Weather {
         this.city = city;
     }
 
-    public Integer getTemperature() {
+    public Double getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(Integer temperature) {
+    public void setTemperature(Double temperature) {
         this.temperature = temperature;
     }
